requirements.txt
astroid==2.4.2
Django==2.2.20
djangorestframework==3.12.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.7.0
pylint==2.6.2
python-gnupg==0.4.7
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
wrapt==1.12.1
